+++ finance/views.py
@@ -1,5 +1,5 @@
 from django.contrib import messages
-from django.urls.base import reverse_lazy
+from django.urls.base import reverse_lazy, reverse
 from django.views.decorators.clickjacking import xframe_options_sameorigin
 from employee.forms import signPaperworkForm
 from finance.estimate_data_utils import calcuateSubcontractedCost, calculateGigCost